rename files and remove unnecessary imports
